% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{are_credentials_for_db_available}
\alias{are_credentials_for_db_available}
\title{Checks that stored  username / password exist for a given database.}
\usage{
are_credentials_for_db_available(database)
}
\arguments{
\item{database}{the database identifier}
}
\value{
\code{TRUE} if stored credentials exist for the given database
}
\description{
Checks that stored  username / password exist for a given database.
}
