% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\docType{data}
\name{BUOY_DATA}
\alias{BUOY_DATA}
\title{The constants holding the name of the 3_BU database}
\format{
An object of class \code{character} of length 1.
}
\usage{
BUOY_DATA
}
\description{
The constants holding the name of the 3_BU database
}
\keyword{datasets}
