% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{DB_BUOYS}
\alias{DB_BUOYS}
\title{Connects to an instance of \code{\link{BUOY_DATA}} on a given server machine using a trusted connection}
\usage{
DB_BUOYS(
  server = get_default_db_server(),
  username = get_username_for_db(BUOY_DATA),
  password = get_password_for_db(BUOY_DATA)
)
}
\arguments{
\item{server}{The server name / IP address (defaults to \code{\link{SERVER_DEFAULT}})}
}
\value{
An ODBC connection to \code{\link{BUOY_DATA}} on \code{server}
}
\description{
Connects to an instance of \code{\link{BUOY_DATA}} on a given server machine using a trusted connection
}
