% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{DB_CONNECT_TO}
\alias{DB_CONNECT_TO}
\title{Connects to a SQL Server database on a given server machine using a trusted connection}
\usage{
DB_CONNECT_TO(
  server = get_default_db_server(),
  database,
  username = NA,
  password = NA,
  database_encoding = "CP1252",
  client_charset = "UTF-8"
)
}
\arguments{
\item{server}{The server name / IP address (defaults to \code{\link{SERVER_DEFAULT}})}

\item{database}{The database name}

\item{username}{The user name}

\item{password}{The password}

\item{database_encoding}{The encoding set for the specific database (defaults to CP1252)}

\item{client_charset}{The character set used by the client (defaults to UTF-8)}
}
\value{
An ODBC connection to \code{database} on \code{server}
}
\description{
Connects to a SQL Server database on a given server machine using a trusted connection
}
\examples{
DB_CONNECT_TO("localhost", "IOTDB")
DB_CONNECT_TO("IOTDB")
}
