% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{db_debug_connections}
\alias{db_debug_connections}
\title{Sets the 'debug' mode for all DB connections}
\usage{
db_debug_connections(mode = NA)
}
\description{
Sets the 'debug' mode for all DB connections
}
