% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{DB_E_RAV}
\alias{DB_E_RAV}
\title{Connects to an instance of the \code{\link{E_RAV}} on a given server}
\usage{
DB_E_RAV(
  server = Sys.getenv("RAV_NEW_DB_SERVER"),
  port = Sys.getenv("RAV_NEW_DB_PORT"),
  database = E_RAV,
  username = Sys.getenv("RAV_NEW_DB_USER"),
  password = Sys.getenv("RAV_NEW_DB_PASSWORD")
)
}
\arguments{
\item{server}{The server name / IP address}

\item{port}{The server port}

\item{database}{The database name (defaults to \code{\link{E_RAV}})}

\item{username}{The username}

\item{password}{The password}
}
\description{
Connects to an instance of the \code{\link{E_RAV}} on a given server
}
