% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{DB_RAV}
\alias{DB_RAV}
\title{Connects to an instance of \code{\link{IOTCVESSELS}} (containing the RAV and the AVL) on a given server machine using a trusted connection}
\usage{
DB_RAV(
  server = get_default_db_server(),
  username = get_username_for_db(IOTCVESSELS),
  password = get_password_for_db(IOTCVESSELS)
)
}
\arguments{
\item{server}{The server name / IP address (defaults to \code{\link{SERVER_DEFAULT}})}

\item{username}{The username (defaults to the standard one for this specific DB)}

\item{password}{The password (defaults to the standard one for this specific DB)}
}
\value{
An ODBC connection to \code{\link{IOTCVESSELS}} on \code{server}
}
\description{
Connects to an instance of \code{\link{IOTCVESSELS}} (containing the RAV and the AVL) on a given server machine using a trusted connection
}
