% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{DB_RAV_NEW}
\alias{DB_RAV_NEW}
\title{Connects to an instance of the new RAV database o a given server}
\usage{
DB_RAV_NEW(server, port, database = "rav_export", username, password)
}
\arguments{
\item{server}{The server name / IP address}

\item{port}{The server port}

\item{database}{The database name}

\item{username}{The username}

\item{password}{The password}
}
\description{
Connects to an instance of the new RAV database o a given server
}
