% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\docType{data}
\name{E_RAV}
\alias{E_RAV}
\title{The constants holding the name of the e-RAV database}
\format{
An object of class \code{character} of length 1.
}
\usage{
E_RAV
}
\description{
The constants holding the name of the e-RAV database
}
\keyword{datasets}
