% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{get_credentials_for_db}
\alias{get_credentials_for_db}
\title{Retrieves the default username / password for a given database.}
\usage{
get_credentials_for_db(database)
}
\arguments{
\item{database}{the database identifier}
}
\value{
the stored credentials for the database (if any)
}
\description{
Retrieves the default username / password for a given database.
}
