% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{get_default_db_server}
\alias{get_default_db_server}
\title{Gets the default server name / IP address for all ODBC request to standard IOTC databases}
\usage{
get_default_db_server()
}
\description{
Gets the default server name / IP address for all ODBC request to standard IOTC databases
}
