% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{get_password_for_db}
\alias{get_password_for_db}
\title{Retrieves the default password for a given database.}
\usage{
get_password_for_db(cache = CREDENTIALS_CACHE, database)
}
\arguments{
\item{cache}{used to store credentials}

\item{database}{the database identifier}
}
\value{
the stored password for the database (if any) or \code{NA}
}
\description{
Retrieves the default password for a given database.
}
