% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{get_username_for_db}
\alias{get_username_for_db}
\title{Retrieves the default username for a given database.}
\usage{
get_username_for_db(database)
}
\arguments{
\item{database}{the database identifier}
}
\value{
the stored username for the database (if any) or \code{NA}
}
\description{
Retrieves the default username for a given database.
}
