% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_dbi_handler.R
\name{getDefaultDBIHandler}
\alias{getDefaultDBIHandler}
\title{getDefaultDBIHandler}
\usage{
getDefaultDBIHandler()
}
\description{
\code{getDefaultDBI} allows to get a default DBI Handler
}
\examples{
  getDefaultDBIHandler()

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
