% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\docType{data}
\name{IOTC_MASTER}
\alias{IOTC_MASTER}
\title{The constants holding the name of the IOTC_MASTER database}
\format{
An object of class \code{character} of length 1.
}
\usage{
IOTC_MASTER
}
\description{
The constants holding the name of the IOTC_MASTER database
}
\keyword{datasets}
