% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\docType{data}
\name{IOTCVESSELS}
\alias{IOTCVESSELS}
\title{The constants holding the name of the IOTCVessels (RAV) database}
\format{
An object of class \code{character} of length 1.
}
\usage{
IOTCVESSELS
}
\description{
The constants holding the name of the IOTCVessels (RAV) database
}
\keyword{datasets}
