% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\docType{data}
\name{ROS}
\alias{ROS}
\title{The constants holding the name of the ROS database}
\format{
An object of class \code{character} of length 1.
}
\usage{
ROS
}
\description{
The constants holding the name of the ROS database
}
\keyword{datasets}
