% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\docType{data}
\name{SERVER_DEFAULT}
\alias{SERVER_DEFAULT}
\title{The constants holding the name of the main IOTC DB Server}
\format{
An object of class \code{character} of length 1.
}
\usage{
SERVER_DEFAULT
}
\description{
The constants holding the name of the main IOTC DB Server
}
\keyword{datasets}
