% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{set_credentials_for_db}
\alias{set_credentials_for_db}
\title{Sets the default username / password for a given database.}
\usage{
set_credentials_for_db(database, username, password)
}
\arguments{
\item{database}{the database identifier}

\item{username}{the username}

\item{password}{the password}
}
\description{
Sets the default username / password for a given database.
}
