% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_dbi_handler.R
\name{setDefaultDBIHandler}
\alias{setDefaultDBIHandler}
\title{setDefaultDBIHandler}
\usage{
setDefaultDBIHandler(handler)
}
\arguments{
\item{handler}{a DB connection handler (function)}
}
\description{
\code{setDefaultDBIHandler} allows to set the default DBI handler (function)
}
\examples{
  setDefaultDBIHandler(DB_IOTDB)

}
\author{
Emmanuel Blondel, \email{emmanuel.blondel1@gmail.com}
}
