% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\docType{data}
\name{WP_CE_RAISED}
\alias{WP_CE_RAISED}
\title{The constants holding the name of the WP_CE_raised database}
\format{
An object of class \code{character} of length 1.
}
\usage{
WP_CE_RAISED
}
\description{
The constants holding the name of the WP_CE_raised database
}
\keyword{datasets}
