% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{DB_CONNECT_TO}
\alias{DB_CONNECT_TO}
\title{Connects to a SQL Server database on a given server machine using a trusted connection}
\usage{
DB_CONNECT_TO(
  server = get_default_db_server(),
  database,
  username = NA,
  password = NA
)
}
\arguments{
\item{server}{The server name / IP address (defaults to \code{\link{SERVER_DEFAULT}})}

\item{database}{The database name}
}
\value{
An ODBC connection to \code{database} on \code{server}
}
\description{
Connects to a SQL Server database on a given server machine using a trusted connection
}
\examples{
DB_CONNECT_TO("localhost", "IOTDB")
DB_CONNECT_TO("IOTDB")
}
