% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotc_core_db_connections_main.R
\name{set_default_db_server}
\alias{set_default_db_server}
\title{Sets the default server name / IP address for all ODBC request to standard IOTC databases}
\usage{
set_default_db_server(cache = SERVER_CACHE, server = SERVER_DEFAULT)
}
\arguments{
\item{cache}{used to store server}

\item{server}{A server name / IP address (defaults to \code{\link{SERVER_DEFAULT}})}
}
\description{
Sets the default server name / IP address for all ODBC request to standard IOTC databases
}
